code generator agent/requirements.txt
transformers 
langgraph 
langchain 
langchain-huggingface 
python-dotenv 
sentence-transformers 
streamlit 
optree>=0.13.0
openai
